+++ zebra-chain/src/sapling/tests.rs
@@ -1,0 +1,1 @@
+mod arbitrary;
